MarfBOT 0.6!
* Fixed Buienradar.

The way it's done is a dirty little fix, It downloads the file from Buienradar, and it saves it in the /var/www/html/MarfBOT Directory as weer.gif
Now can discord see it as a real gif and it is working in chat. There's a way to send a local file but i did this for now because reading the documentation is a drama.

Signed-off-by: Marvin Ferwerda <marfie1997@ziggo.nl>