Change readme.md for deprecation of this project.
This is also the very last commit i will do for MarfBOT.
it was fun, and it was kind of my baby because i cared about it alot, Very funny for a simple stupid Discord Bot. but oh well.
As MarfBOT broke so many times, in october 2019 it was the last time that i did the uncommited changes to fix it because i was tired of it and i didnt use and other people use the bot anymore.

Goodbye MarfBOT.
Signed-off-by: Marf <git-no-reply@marfprojects.nl>